refactor: rename internal/clustername to internal/oidcissuer
The package no longer resolves a friendly cluster name -- it extracts
the OIDC issuer claim as an opaque provider identity. Name the package
for what it does.
