+++ internal/oidcissuer/resolver.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package clustername
+package oidcissuer
 
 import (
 	"encoding/base64"
